Extract decorators from index.ts to the decorators directory
