+++ decorators/event_emitter.ts
@@ -1,0 +1,11 @@
+export class EventEmitter {
+    private listeners: Array<Function> = [];
+
+    emit(event: any = null) {
+        this.listeners.forEach(callback => callback.call(null, event));
+    }
+
+    subscribe(callback: Function) {
+        this.listeners.push(callback);
+    }
+}
