+++ decorators/inject/index.ts
@@ -1,0 +1,11 @@
+import {IComponentClass} from "../component/interfaces";
+
+export function Inject(dependencyName: string): ParameterDecorator {
+    return function (target: IComponentClass, property: string, parameterIndex: number) {
+        if (!target.$inject) {
+            target.$inject = [];
+        }
+
+        target.$inject[parameterIndex] = dependencyName;
+    };
+}
