+++ test/input.test.ts
@@ -1,11 +1,11 @@
 import "reflect-metadata";
-import {Input} from "../core";
+import {Input} from "../decorators/input";
 
 class Test {
     @Input("prop2") prop: string;
 }
 
-describe("Input decorator", function() {
+describe("input decorator", function() {
     let instance: any;
 
     beforeEach(function() {
